% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsens.R
\name{gsensX}
\alias{gsensX}
\title{Adjusting for genetic confounding using PGS for the exposure
Adjusting for genetic confounding in exposure-outcome associations using the polygenic score for the exposure}
\usage{
gsensX(rxy, rgx, rgy, n, h2, constrain = NULL, print = FALSE)
}
\arguments{
\item{rxy}{The observed phenotypic correlation between exposure X and outcome Y.}

\item{rgx}{The observed correlation between the polygenic score for X and exposure X.}

\item{rgy}{The observed correlation between the polygenic score for X and outcome Y.}

\item{n}{Sample size}

\item{h2}{The additive genetic variance explained in exposure X under the scenario of interest.}

\item{constrain}{Optional. Argument to constrain model parameters (default = NULL).}

\item{print}{Optional. Enables the examination of model parameters (default = FALSE).}
}
\value{
Estimates for the adjusted exposure-outcome associations, exposure-mediated genetic effects,
genetic confounding and genetic overlap.
}
\description{
Adjusting for genetic confounding using PGS for the exposure
Adjusting for genetic confounding in exposure-outcome associations using the polygenic score for the exposure
}
\references{
Pingault, J.-B., O’Reilly, P. F., Schoeler, T., Ploubidis, G. B., Rijsdijk, F., & Dudbridge, F. (2018).
Using genetic data to strengthen causal inference in observational research. Nature Reviews Genetics, 19(9), 566–580.
https://doi.org/10.1038/s41576-018-0020-3

Pingault, J. B., Rijsdijk, F., Schoeler, T., Choi, S. W., Selzam, S., Krapohl, E., ... & Dudbridge, F. (2021).
Genetic sensitivity analysis: Adjusting for genetic confounding in epidemiological associations. PLoS genetics, 17(6), e1009590.
https://doi.org/10.1371/journal.pgen.1009590
}
\author{
Jean-Baptiste Pingault, Tabea Schoeler & Frank Dudbridge
}
