% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsens.R
\name{gsensXY}
\alias{gsensXY}
\title{Adjusting for genetic confounding using PGS for the exposure and the outcome}
\usage{
gsensXY(
  rxy,
  rg1x,
  rg2y,
  rg1y,
  rg2x,
  rg1g2,
  n,
  h2.x,
  h2.y,
  constrain = NULL,
  print = FALSE
)
}
\arguments{
\item{rxy}{The observed phenotypic correlation between exposure X and outcome Y.}

\item{rg1x}{The observed correlation between the polygenic score for X and exposure X.}

\item{rg2y}{The observed correlation between the polygenic score for X and outcome Y.}

\item{rg1y}{The correlation between the outcome Y and the observed polygenic score for X.}

\item{rg2x}{The correlation between the exposure X and the observed polygenic score for Y.}

\item{rg1g2}{The correlation between the two observed polygenic scores.}

\item{n}{Sample size}

\item{h2.x}{The additive genetic variance explained in exposure X under the scenario of interest.}

\item{h2.y}{The additive genetic variance explained in exposure Y under the scenario of interest.}

\item{constrain}{Optional. Argument to constrain model parameters (default = NULL).}

\item{print}{Optional. Enables the examination of model parameters (default = FALSE).}
}
\value{
Estimates for the adjusted exposure-outcome association,
genetic confounding and the total effect.
}
\description{
Adjusting for genetic confounding in exposure-outcome associations using the polygenic score for the exposure and the outcome.
}
\references{
Pingault, J.-B., O’Reilly, P. F., Schoeler, T., Ploubidis, G. B., Rijsdijk, F., & Dudbridge, F. (2018).
Using genetic data to strengthen causal inference in observational research. Nature Reviews Genetics, 19(9), 566–580.
https://doi.org/10.1038/s41576-018-0020-3

Pingault, J. B., Rijsdijk, F., Schoeler, T., Choi, S. W., Selzam, S., Krapohl, E., ... & Dudbridge, F. (2021).
Genetic sensitivity analysis: Adjusting for genetic confounding in epidemiological associations. PLoS genetics, 17(6), e1009590.
https://doi.org/10.1371/journal.pgen.1009590
}
\author{
Jean-Baptiste Pingault, Tabea Schoeler & Frank Dudbridge
}
