% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsens.R
\name{gsensY}
\alias{gsensY}
\title{Adjusting for genetic confounding using PGS for the outcome}
\usage{
gsensY(
  data,
  sample.nobs = NULL,
  h2,
  estimator = "GLS",
  se = "standard",
  boot.ci.type = "perc",
  fmi = F,
  print = "summary",
  ...
)
}
\arguments{
\item{data}{Either a data frame of raw data or a covariance/correlation matrix. 
Must contain the variables G (PGS for Y), Y (outcome) and X (exposure) or instead X1, X2, ..., Xi if multiple exposures are used.
If data is a cov/cor matrix, the additional argument sample.nobs is required.}

\item{sample.nobs}{Optional. Number of observations used for the data.
Only necessary if input is a covariance or correlation matrix.}

\item{h2}{Heritability estimate of the outcome (Y).
Can be chosen to be any external value, e.g. SNP- or twin-heritability estimates.}

\item{estimator}{Optional. Type of estimator used.
Can be any estimator for continuous data implemented in lavaan, e.g. "ML" or "GLS" (default = "GLS").}

\item{se}{Optional. Method to compute standard errors, can be any method implemented in lavaan,
e.g. "robust.sem" or "bootstrap". If se = "bootstrap", bootstrapping *cannot* be used for estimation of confidence intervals,
which is however strongly recommended (default = "standard").}

\item{boot.ci.type}{Bootstrapping method used to compute the confidence intervals,
can be any method implemented in lavaan. Default is "perc".}

\item{fmi}{Optional. If full-information maximum likelihood estimation is desired (estimator = "ML", missing "ML"),
and fmi = TRUE, missingness is reported (default = FALSE).}

\item{print}{Optional. Can be one of c("exposure", "mediation", "confounding", "overlap", "all" or "summary").
If print = "all", all relevant parameters will be printed.
If print = "summary", the lavaan output of the summary() function will be printed (default = "summary").}

\item{...}{Additional arguments passed from lavaan, including 'bootstrap' (number of bootstraps for CIs, default = 1000)
and 'constraints' (constraints on the model).}
}
\value{
Estimates for the adjusted exposure-outcome associations, exposure-mediated genetic effects,
genetic confounding and genetic overlap.
}
\description{
Adjusting for genetic confounding in exposure-outcome associations using the polygenic score for the outcome.
This is the recommended function for most scenarios, and the only function that has been extended to the multiple exposure case.
}
\examples{
df <- data.frame(G, X1, X2, Y) 
gsensY(df, h2 = 0.5);
}
\references{
Frach, L., Rijsdijk, F., Dudbridge, F. & Pingault, J. B. (in preparation).
Adjusting for genetic confounding using polygenic scores within structural equation models

Pingault, J. B., Rijsdijk, F., Schoeler, T., Choi, S. W., Selzam, S., Krapohl, E., ... & Dudbridge, F. (2021).
Genetic sensitivity analysis: Adjusting for genetic confounding in epidemiological associations. PLoS genetics, 17(6), e1009590. https://doi.org/10.1371/journal.pgen.1009590
}
\author{
Leonard Frach & Jean-Baptiste Pingault
}
