% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsens.R
\name{gsensY}
\alias{gsensY}
\title{Adjusting for genetic confounding using PGS for the outcome}
\usage{
gsensY(
  data = NULL,
  sample.cov = NULL,
  h2,
  exposures,
  outcome,
  pgs,
  model = NULL,
  loading = FALSE,
  lg2 = NULL,
  ...
)
}
\arguments{
\item{data}{Optional. Used when input contains raw data (data frame).}

\item{sample.cov}{Optional. If input is a covariance or correlation matrix, the lavaan arguments \code{sample.cov = df_cov} and \code{sample.nobs = n} are required, where \code{df_cov} should be the name of your input matrix and \code{n} the sample size.}

\item{h2}{Heritability estimate of the outcome (Y).
Can be chosen to be any external value, e.g. SNP- or twin-heritability estimates.}

\item{exposures}{Vector of variable name(s) of the exposure(s).
Example: \code{exposures = c("x1", "x2")}}

\item{outcome}{Name of the outcome variable.}

\item{pgs}{Name of the polygenic score variable (pgs corresponding to the outcome).}

\item{model}{Optional. If additional variables beyond exposures, PGS and outcome are used, e.g., measured confounders,
these additional relations can be specified and will be added to the GsensY model. Example: \code{model = c("X1 ~ Z", "X2 ~ Z")}.
If \code{model} is not specified, the standard GsensY model will be estimated.}

\item{...}{Additional arguments passed from lavaan, including \code{se} (estimation method for the standard errors),
\code{estimator} (estimator used for model, default is ML), \code{bootstrap} (number of bootstraps for CIs, default = 1000),
\code{sample.nobs} (Number of observations for estimation using summary data, not recommended), and more.
See the lavaan documentation for details, e.g., \code{\link[lavaan:lavaan]{lavaan::lavaan()}} or \code{\link[lavaan:lavOptions]{lavaan::lavOptions()}}}
}
\value{
The Gsens model output will be returned as a lavaan object.
For example, the \code{summary()} or \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}} functions can be used for more detailed outputs, e.g. for standardized estimates.
}
\description{
Adjusting for genetic confounding in exposure--outcome associations using the polygenic score for the outcome.
This is the recommended function for most scenarios, and the only function that has been extended to the multiple exposure case.
}
\examples{
\dontrun{
df <- data.frame(X1, X2, X3, Y, PGS_outcome)
df_cov <- cov(df)
gsens_out <- gsensY(data = df, h2 = 0.5, exposures = c("X1", "X2", "X3"), outcome = "Y", pgs = "PGS_outcome")
gsens_out_cov <- gsensY(sample.cov = df_cov, sample.nobs = 5000, h2 = 0.5, exposures = c("X1", "X2", "X3"), outcome = "Y", pgs = "PGS_outcome")

## print GsensY results
gsens_out@external$gsensY
## any other standard lavaan output also available
summary(gsens_out)
}

}
\references{
Frach, L., Rijsdijk, F., Dudbridge, F. & Pingault, J. B. (in preparation).
Adjusting for genetic confounding using polygenic scores within structural equation models.

Pingault, J. B., Rijsdijk, F., Schoeler, T., Choi, S. W., Selzam, S., Krapohl, E., ... & Dudbridge, F. (2021).
Genetic sensitivity analysis: Adjusting for genetic confounding in epidemiological associations.
\emph{PLoS genetics, 17}(6), e1009590. \doi{10.1371/journal.pgen.1009590}
}
\author{
Leonard Frach & Jean-Baptiste Pingault
}
